Only produce pixel backplane for valid pixels
Previously, mapped backplane pixel coordinates were extrapolated beyond the bounds of the image (e.g. a 30x30 image could have a backplane map with pixel coordinate values >30). This was (a) misleading and (b) broke some of the mapping interpolation, so the new version checks pixel coordinates are valid before adding them to the map.