Add decorator to adjust alt for mapping
Maps may or may not have the alt argument set, so use a new subclassed context manager that can be reentrant, and a decorator that can consume map_kwargs. This new decorator/context manager will then adjust the altitude at the top of the stack, and nested calls won't make any further adjustment.