Reduce the severity of the "as is" warning in the docs
The planetmapper API has been stable for a few years now, so the "active development" warning in the code was a bit too severe. Therefore, downgrade it to the more standard "code provided as is... use at your own risk" type warning more common in open source code (reflecting the boilerplate language in the MIT license).