Add .python-version for dependabot
Add a .python-version file to indicate to dependabot that we would prefer to run with a more recent version of Python
(https://github.com/dependabot/dependabot-core/issues/1455#issuecomment-545389290).
Previously, dependabot was using the oldest version of Python (3.10) compatble with the setup.py python_requires (>=3.10). This meant that it was unable to update the scipy dependencies to the latest versions, as these require >=3.11. As our requirements.txt uses upper bounds, we want dependabot to run with a recent Python version to ensure the upper bound is up-to-date with the latest package versions.